refactor: decompose CoWShedWidget
