+++ apps/cowswap-frontend/src/modules/cowShed/containers/CoWShedWidget/styled.tsx
@@ -3,6 +3,8 @@ import { UI } from '@cowprotocol/ui'
 import styled from 'styled-components/macro'
 import { WIDGET_MAX_WIDTH } from 'theme'
 
+import { AddressLink } from 'common/pure/AddressLink'
+
 export const Wrapper = styled.div`
   width: 100%;
   max-width: ${WIDGET_MAX_WIDTH.swap};
@@ -25,37 +27,18 @@ export const Wrapper = styled.div`
   }
 `
 
-export const ProxyInfo = styled.div`
+export const AddressLinkStyled = styled(AddressLink)`
+  color: inherit;
+  width: 100%;
+  font-size: 100%;
+  background: var(${UI.COLOR_PAPER});
+  border-radius: 16px;
+  padding: 10px;
   display: flex;
-  flex-flow: column wrap;
-  gap: 10px;
-  margin: 0;
-  text-align: center;
-  font-size: 15px;
-
-  > h4 {
-    font-size: 16px;
-    font-weight: 600;
-    margin: 0 auto;
-  }
-
-  > a {
-    color: inherit;
-    width: 100%;
-  }
-
-  > a > span {
-    font-size: 100%;
-    background: var(${UI.COLOR_PAPER});
-    border-radius: 16px;
-    padding: 10px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    width: 100%;
-    margin: 0 auto;
-    word-break: break-all;
-  }
+  align-items: center;
+  justify-content: center;
+  margin: 0 auto;
+  word-break: break-all;
 `
 
 export const Content = styled.div`
